refactor: mv code around
