+++ packages/axios-extension/test/auth/reentrance-ticket/abap-system.test.ts
@@ -1,4 +1,4 @@
-import { ABAPSystem } from '@src/auth/abap-system';
+import { ABAPSystem } from '@src/auth/reentrance-ticket/abap-system';
 
 describe('ABAPSystem', () => {
     describe('uiHostname()', () => {
